move cli into dedicated directory
